+++ crypto/secp256k1.go
@@ -33,7 +33,7 @@ func GenerateKey() (PrivKeySecp256k1, error) {
 // PubKey returns the ECDSA private key's public key.
 func (privkey PrivKeySecp256k1) PubKey() tmcrypto.PubKey {
 	ecdsaPKey := privkey.ToECDSA()
-	return PubKeySecp256k1(ethcrypto.FromECDSAPub(&ecdsaPKey.PublicKey))
+	return PubKeySecp256k1(ethcrypto.CompressPubkey(&ecdsaPKey.PublicKey))
 }
 
 // Bytes returns the raw ECDSA private key bytes.
@@ -58,8 +58,12 @@ func (privkey PrivKeySecp256k1) Equals(other tmcrypto.PrivKey) bool {
 }
 
 // ToECDSA returns the ECDSA private key as a reference to ecdsa.PrivateKey type.
+// The function will panic if the private key is invalid.
 func (privkey PrivKeySecp256k1) ToECDSA() *ecdsa.PrivateKey {
-	key, _ := ethcrypto.ToECDSA(privkey)
+	key, err := ethcrypto.ToECDSA(privkey)
+	if err != nil {
+		panic(err)
+	}
 	return key
 }
 
@@ -68,17 +72,23 @@ func (privkey PrivKeySecp256k1) ToECDSA() *ecdsa.PrivateKey {
 
 var _ tmcrypto.PubKey = (*PubKeySecp256k1)(nil)
 
-// PubKeySecp256k1 defines a type alias for an ecdsa.PublicKey that implements
-// Tendermint's PubKey interface.
+// PubKeySecp256k1 defines a type alias for an ecdsa.PublicKey that implements Tendermint's PubKey
+// interface. It represents the 33-byte compressed public key format.
 type PubKeySecp256k1 []byte
 
 // Address returns the address of the ECDSA public key.
+// The function will panic if the public key is invalid.
 func (key PubKeySecp256k1) Address() tmcrypto.Address {
-	pubk, _ := ethcrypto.UnmarshalPubkey(key)
+	pubk, err := ethcrypto.DecompressPubkey(key)
+	if err != nil {
+		panic(err)
+	}
+
 	return tmcrypto.Address(ethcrypto.PubkeyToAddress(*pubk).Bytes())
 }
 
 // Bytes returns the raw bytes of the ECDSA public key.
+// The function panics if the key cannot be marshaled to bytes.
 func (key PubKeySecp256k1) Bytes() []byte {
 	bz, err := CryptoCodec.MarshalBinaryBare(key)
 	if err != nil {