+++ types/account.go
@@ -11,8 +11,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/auth/exported"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 
-	tmamino "github.com/tendermint/tendermint/crypto/encoding/amino"
-
 	ethcmn "github.com/ethereum/go-ethereum/common"
 	ethcrypto "github.com/ethereum/go-ethereum/crypto"
 )
@@ -79,7 +77,7 @@ func (acc *EthAccount) SetBalance(amt sdk.Int) {
 type ethermintAccountPretty struct {
 	Address       sdk.AccAddress `json:"address" yaml:"address"`
 	Coins         sdk.Coins      `json:"coins" yaml:"coins"`
-	PubKey        []byte         `json:"public_key" yaml:"public_key"`
+	PubKey        string         `json:"public_key" yaml:"public_key"`
 	AccountNumber uint64         `json:"account_number" yaml:"account_number"`
 	Sequence      uint64         `json:"sequence" yaml:"sequence"`
 	CodeHash      string         `json:"code_hash" yaml:"code_hash"`
@@ -95,8 +93,13 @@ func (acc EthAccount) MarshalYAML() (interface{}, error) {
 		CodeHash:      ethcmn.Bytes2Hex(acc.CodeHash),
 	}
 
+	var err error
+
 	if acc.PubKey != nil {
-		alias.PubKey = acc.PubKey.Bytes()
+		alias.PubKey, err = sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeAccPub, acc.PubKey)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	bz, err := yaml.Marshal(alias)
@@ -117,8 +120,13 @@ func (acc EthAccount) MarshalJSON() ([]byte, error) {
 		CodeHash:      ethcmn.Bytes2Hex(acc.CodeHash),
 	}
 
+	var err error
+
 	if acc.PubKey != nil {
-		alias.PubKey = acc.PubKey.Bytes()
+		alias.PubKey, err = sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeAccPub, acc.PubKey)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	return json.Marshal(alias)
@@ -126,26 +134,34 @@ func (acc EthAccount) MarshalJSON() ([]byte, error) {
 
 // UnmarshalJSON unmarshals raw JSON bytes into an EthAccount.
 func (acc *EthAccount) UnmarshalJSON(bz []byte) error {
-	acc.BaseAccount = &authtypes.BaseAccount{}
-	var alias ethermintAccountPretty
+	var (
+		alias ethermintAccountPretty
+		err   error
+	)
+
 	if err := json.Unmarshal(bz, &alias); err != nil {
 		return err
 	}
 
-	if alias.PubKey != nil {
-		pubKey, err := tmamino.PubKeyFromBytes(alias.PubKey)
+	acc.BaseAccount = &authtypes.BaseAccount{
+		Coins:         alias.Coins,
+		Address:       alias.Address,
+		AccountNumber: alias.AccountNumber,
+		Sequence:      alias.Sequence,
+	}
+	acc.CodeHash = ethcmn.Hex2Bytes(alias.CodeHash)
+
+	if alias.PubKey != "" {
+		acc.BaseAccount.PubKey, err = sdk.GetPubKeyFromBech32(sdk.Bech32PubKeyTypeAccPub, alias.PubKey)
 		if err != nil {
 			return err
 		}
-
-		acc.BaseAccount.PubKey = pubKey
 	}
-
-	acc.BaseAccount.Coins = alias.Coins
-	acc.BaseAccount.Address = alias.Address
-	acc.BaseAccount.AccountNumber = alias.AccountNumber
-	acc.BaseAccount.Sequence = alias.Sequence
-	acc.CodeHash = ethcmn.Hex2Bytes(alias.CodeHash)
-
 	return nil
+}
+
+// String implements the fmt.Stringer interface
+func (acc EthAccount) String() string {
+	out, _ := yaml.Marshal(acc)
+	return string(out)
 }